Implement feature-flag module
This module now delegates a request to a feature release directory. Delegation
happens driven by a manifest file called `flags.json`.

It currently does not retrieve any information from the request to map a it to
a defined feature flag, so it always defaults to `master`. For now, it only
supports arrays of account IDs for the FF condition but this might change in
the future in favor of some more sophisticated fully dynamic callback.